Update phonebook functionality: modify person data structure, enhance add and delete person features with notifications, and improve error handling

+++ part2/phonebook/src/App.jsx
@@ -3,12 +3,17 @@ import PersonForm from "./components/PersonForm";
 import SearchFilter from "./components/SearchFilter";
 import Persons from "./components/Persons";
 import personService from './services/persons'
+import Notification from "./components/Notification";
+import './index.css'
+
 
 const App = () => {
   const [persons, setPersons] = useState([]);
   const [newName, setNewName] = useState("");
   const [newNumber, setNewNumber] = useState("");
   const [filterWord, setFilterWord] = useState("");
+  const [errorMessage, setErrorMessage] = useState("")
+  const [isGreen, setIsGreen] = useState(true)
 
   useEffect(() => {
     personService
@@ -18,32 +23,41 @@ const App = () => {
       });
   }, []);
 
-    const addPerson = (e) => {
-        e.preventDefault();
-        if (persons.find((person) => person.name === newName)) {
-            alert(`${newName} is already added to phonebook`);
-        } else {
-            const nameObject = {
-                name: newName,
-                id: persons.length + 1,
-                number: newNumber,
-            };
-            personService
-                .create(nameObject)
-                .then(initialPerson => {
-                    setPersons(persons.concat(initialPerson))
-                    setNewName("");
-                })
-        }
-    };
-
-    const deletePerson = (id) => {
-        personService
-            .deletePerson(id)
-            .then(() => {
-                setPersons(persons.filter(n => n.id !== id))
-            })
+  const addPerson = (e) => {
+    e.preventDefault();
+    if (persons.find((person) => person.name === newName)) {
+      alert(`${newName} is already added to phonebook`);
+    } else {
+      const nameObject = {
+        name: newName,
+        number: newNumber,
+      };
+      personService
+        .create(nameObject)
+        .then(initialPerson => {
+          setIsGreen(true)
+          setPersons(persons.concat(initialPerson))
+          setNewName("");
+          setErrorMessage(`${nameObject.name} added in phonebook`)
+        })
     }
+  };
+
+  const deletePerson = (id) => {
+    const person = persons.find(n => n.id === id)
+    if (window.confirm(`Delete ${person.name}?`)) {
+      personService
+        .deletePerson(id)
+        .then(() => {
+
+          setPersons(persons.filter(n => n.id !== id))
+        })
+        .catch(error => {
+          setIsGreen(false)
+          setErrorMessage(`${person.id} has already removed from server`)
+        })
+    }
+  }
 
   const handlePersonChange = (e) => {
     console.log(e.target.value);
@@ -61,6 +75,7 @@ const App = () => {
   return (
     <div>
       <h2>Phonebook</h2>
+      <Notification message={errorMessage} isGreen={isGreen} />
       <SearchFilter value={filterWord} onChange={handleFilterChange} />
 
       <h3>Add new</h3>
